add integration test crate
