+++ .travis.yml
@@ -43,7 +43,7 @@ matrix:
       rust: stable
 
     - rust: stable
-      env: MODULES="compatibility"
+      env: MODULES="compatibility-tests"
            TDNS_BIND_PATH="../../bind-9.11.0-P1/bin/named/named"
            OPTIONS="--no-default-features --features=bind"
 
@@ -52,7 +52,7 @@ matrix:
 
 before_install:
   - if [[ "$TRAVIS_OS_NAME" == "linux" ]]; then scripts/install_openssl_deb.sh ; fi
-  - if [[ "$MODULES" == "compatibility" ]]; then scripts/install_bind.sh ; fi
+  - if [[ "$MODULES" == "compatibility-tests" ]]; then scripts/install_bind.sh ; fi
 script: scripts/run_tests.sh
 after_success: scripts/run_kcov.sh
 notifications:
