+++ compatibility-tests/tests/conf/bind-example.conf
@@ -42,7 +42,7 @@ zone "0.in-addr.arpa" {
 
 zone "example.com" {
   type master;
-  file "../../compatibility/tests/conf/bind-example.com.zone";
+  file "../../compatibility-tests/tests/conf/bind-example.com.zone";
   journal "./bind-example.com.zone.jnl";
 
   allow-query { any; };
