+++ integration-tests/Cargo.toml
@@ -1,0 +1,53 @@
+[package]
+name = "trust-dns-integration"
+version = "0.1.0"
+authors = ["Benjamin Fry <benjaminfry@me.com>"]
+
+# A short blurb about the package. This is not rendered in any format when
+# uploaded to crates.io (aka this is not markdown)
+description = """
+TRust-DNS integration testing library.
+"""
+
+# These URLs point to more information about the repository
+documentation = "https://docs.rs/trust-dns"
+homepage = "http://www.trust-dns.org/index.html"
+repository = "https://github.com/bluejekyll/trust-dns"
+
+# This points to a file in the repository (relative to this Cargo.toml). The
+# contents of this file are stored and indexed in the registry.
+readme = "../README.md"
+
+# This is a small list of keywords used to categorize and search for this
+# package.
+keywords = ["DNS", "BIND", "dig", "named", "dnssec"]
+categories = ["network-programming"]
+
+# This is a string description of the license for this package. Currently
+# crates.io will validate the license provided against a whitelist of known
+# license identifiers from http://spdx.org/licenses/. Multiple licenses can
+# be separated with a `/`
+license = "MIT/Apache-2.0"
+
+[badges]
+travis-ci = { repository = "bluejekyll/trust-dns" }
+appveyor = { repository = "bluejekyll/trust-dns", branch = "master", service = "github" }
+
+[lib]
+name = "trust_dns_integration"
+path = "src/lib.rs"
+
+[dependencies]
+chrono = "^0.4"
+data-encoding = "^1.2.0"
+env_logger = "0.4.2"
+log = "^0.3.5"
+futures = "^0.1.6"
+openssl = { version = "^0.9.8", features = ["v102", "v110"] }
+rand = "^0.3"
+rusqlite = { version = "^0.9.5", features = ["bundled"] }
+tokio-core = "^0.1"
+trust-dns = { version = "*", path="../client" }
+trust-dns-resolver = { version = "*", path="../resolver" }
+trust-dns-server = { version = "*", path="../server" }
+
