+++ resolver/README.md
@@ -10,9 +10,9 @@ This library contains implementations for IPv4 (A) and IPv6 (AAAA) resolution, m
 - `/etc/resolv.conf` based configuration on Unix/Posix systems
 - NameServer pools with performance based priority usage
 - Caching of query results
-- NXDOMAIN/NoData caching (negative caching)
+- NxDomain/NoData caching (negative caching)
 - DNSSec validation
-- Generic Record Type Lookup (the Client can be used in the interim)
+- Generic Record Type Lookup
 - *TBD* TLS integration
 
 ## Example
