+++ scripts/run_tests.sh
@@ -1,6 +1,6 @@
 #!/bin/bash -e
 
-MODULES=${MODULES:-"client native-tls resolver rustls server"}
+MODULES=${MODULES:-"client integration-tests native-tls resolver rustls server"}
 CLIENT_OPTIONS=${CLIENT_OPTIONS} # add in all features
 OPTIONS=${OPTIONS}
