+++ server/tests/z_named_tests.rs
@@ -6,7 +6,7 @@ extern crate tokio_core;
 extern crate trust_dns_server;
 extern crate openssl;
 
-mod common;
+mod server_harness;
 
 use std::env;
 use std::fs::File;
@@ -21,7 +21,7 @@ use trust_dns::client::*;
 use trust_dns::tcp::TcpClientStream;
 use trust_dns::tls::TlsClientStream;
 
-use common::server_harness::{named_test_harness, query};
+use server_harness::{named_test_harness, query};
 
 #[test]
 fn test_example_toml_startup() {
@@ -31,14 +31,22 @@ fn test_example_toml_startup() {
 
     named_test_harness("example.toml", |port, _| {
         let mut io_loop = Core::new().unwrap();
-        let addr: SocketAddr = ("127.0.0.1", port).to_socket_addrs().unwrap().next().unwrap();
+        let addr: SocketAddr = ("127.0.0.1", port)
+            .to_socket_addrs()
+            .unwrap()
+            .next()
+            .unwrap();
         let (stream, sender) = TcpClientStream::new(addr, &io_loop.handle());
         let mut client = ClientFuture::new(stream, sender, &io_loop.handle(), None);
 
         assert!(query(&mut io_loop, &mut client));
 
         // just tests that multiple queries work
-        let addr: SocketAddr = ("127.0.0.1", port).to_socket_addrs().unwrap().next().unwrap();
+        let addr: SocketAddr = ("127.0.0.1", port)
+            .to_socket_addrs()
+            .unwrap()
+            .next()
+            .unwrap();
         let (stream, sender) = TcpClientStream::new(addr, &io_loop.handle());
         let mut client = ClientFuture::new(stream, sender, &io_loop.handle(), None);
 
@@ -50,7 +58,11 @@ fn test_example_toml_startup() {
 fn test_ipv4_only_toml_startup() {
     named_test_harness("ipv4_only.toml", |port, _| {
         let mut io_loop = Core::new().unwrap();
-        let addr: SocketAddr = ("127.0.0.1", port).to_socket_addrs().unwrap().next().unwrap();
+        let addr: SocketAddr = ("127.0.0.1", port)
+            .to_socket_addrs()
+            .unwrap()
+            .next()
+            .unwrap();
         let (stream, sender) = TcpClientStream::new(addr, &io_loop.handle());
         let mut client = ClientFuture::new(stream, sender, &io_loop.handle(), None);
 
@@ -97,7 +109,11 @@ fn test_ipv4_only_toml_startup() {
 fn test_ipv4_and_ipv6_toml_startup() {
     named_test_harness("ipv4_and_ipv6.toml", |port, _| {
         let mut io_loop = Core::new().unwrap();
-        let addr: SocketAddr = ("127.0.0.1", port).to_socket_addrs().unwrap().next().unwrap();
+        let addr: SocketAddr = ("127.0.0.1", port)
+            .to_socket_addrs()
+            .unwrap()
+            .next()
+            .unwrap();
         let (stream, sender) = TcpClientStream::new(addr, &io_loop.handle());
         let mut client = ClientFuture::new(stream, sender, &io_loop.handle(), None);
 
@@ -122,13 +138,19 @@ fn test_example_tls_toml_startup() {
         let server_path = env::var("TDNS_SERVER_SRC_ROOT").unwrap_or(".".to_owned());
         println!("using server src path: {}", server_path);
 
-        File::open(&format!("{}/tests/named_test_configs/sec/example.cert", server_path))
-            .unwrap()
+        File::open(&format!(
+            "{}/tests/named_test_configs/sec/example.cert",
+            server_path
+        )).unwrap()
             .read_to_end(&mut cert_der)
             .unwrap();
 
         let mut io_loop = Core::new().unwrap();
-        let addr: SocketAddr = ("127.0.0.1", tls_port).to_socket_addrs().unwrap().next().unwrap();
+        let addr: SocketAddr = ("127.0.0.1", tls_port)
+            .to_socket_addrs()
+            .unwrap()
+            .next()
+            .unwrap();
         let mut tls_conn_builder = TlsClientStream::builder();
         let cert = to_trust_anchor(&cert_der);
         tls_conn_builder.add_ca(cert);
@@ -139,7 +161,11 @@ fn test_example_tls_toml_startup() {
         // ipv4 should succeed
         assert!(query(&mut io_loop, &mut client));
 
-        let addr: SocketAddr = ("127.0.0.1", tls_port).to_socket_addrs().unwrap().next().unwrap();
+        let addr: SocketAddr = ("127.0.0.1", tls_port)
+            .to_socket_addrs()
+            .unwrap()
+            .next()
+            .unwrap();
         let mut tls_conn_builder = TlsClientStream::builder();
         let cert = to_trust_anchor(&cert_der);
         tls_conn_builder.add_ca(cert);
