+++ tests/test.rs
@@ -1,0 +1,4 @@
+#[test]
+fn try_one() {
+	assert!(false)
+}
